requirements.txt
python-dotenv==1.0.1
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.2.1
openpyxl==3.1.2
requests==2.31.0
pdf2text==1.0.0
PyPDF2==3.0.1
google-generativeai==0.4.0 